+++ src/apps/aelin/aelin.module.ts
@@ -5,7 +5,7 @@ import { SynthetixAppModule } from '~apps/synthetix/synthetix.module';
 import { AelinAppDefinition, AELIN_DEFINITION } from './aelin.definition';
 import { AelinContractFactory } from './contracts';
 import { EthereumAelinPoolTokenFetcher } from './ethereum/aelin.pool.token-fetcher';
-import { OptimismAelinBalanceFetcher } from './optimism/aelin.balance-fetcher';
+import { OptimismAelinFarmContractPositionBalanceFetcher } from './optimism/aelin.farm.contract-position-balance-fetcher';
 import { OptimismAelinFarmContractPositionFetcher } from './optimism/aelin.farm.contract-position-fetcher';
 import { OptimismAelinPoolTokenFetcher } from './optimism/aelin.pool.token-fetcher';
 import { OptimismAelinVAelinTokenFetcher } from './optimism/aelin.v-aelin.token-fetcher';
@@ -21,8 +21,8 @@ import { OptimismAelinVAelinTokenFetcher } from './optimism/aelin.v-aelin.token-
     // Optimism
     OptimismAelinPoolTokenFetcher,
     OptimismAelinVAelinTokenFetcher,
-    OptimismAelinBalanceFetcher,
     OptimismAelinFarmContractPositionFetcher,
+    OptimismAelinFarmContractPositionBalanceFetcher,
   ],
 })
 export class AelinAppModule extends AbstractApp() {}
