+++ src/apps/alpha-v1/alpha-v1.module.ts
@@ -3,7 +3,6 @@ import { AbstractApp } from '~app/app.dynamic-module';
 
 import ALPHA_V1_DEFINITION, { AlphaV1AppDefinition } from './alpha-v1.definition';
 import { AlphaV1ContractFactory } from './contracts';
-import { EthereumAlphaV1BalanceFetcher } from './ethereum/alpha-v1.balance-fetcher';
 import { EthereumAlphaV1LendingTokenFetcher } from './ethereum/alpha-v1.lending.token-fetcher';
 import { EthereumAlphaV1TvlFetcher } from './ethereum/alpha-v1.tvl-fetcher';
 
@@ -12,7 +11,6 @@ import { EthereumAlphaV1TvlFetcher } from './ethereum/alpha-v1.tvl-fetcher';
   providers: [
     AlphaV1AppDefinition,
     AlphaV1ContractFactory,
-    EthereumAlphaV1BalanceFetcher,
     EthereumAlphaV1LendingTokenFetcher,
     EthereumAlphaV1TvlFetcher,
   ],
