+++ src/balance/default.balance-afterware.factory.ts
@@ -1,4 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
+import { groupBy, values } from 'lodash';
 
 import { presentBalanceFetcherResponse } from '~app-toolkit/helpers/presentation/balance-fetcher-response.present';
 import { AppService } from '~app/app.service';
@@ -21,13 +22,20 @@ export class DefaultBalanceAfterwareFactory {
       constructor(readonly appService: AppService) {}
 
       async use(balances: PositionBalance[]) {
+        // Build labelled groups by the labels defined in the app definition
         const app = this.appService.getApp(appId);
         const products = Object.values(app.groups).map(group => {
           const groupBalances = balances.filter(v => v.groupId === group.id);
           return { label: group.label, assets: groupBalances };
         });
 
-        return presentBalanceFetcherResponse(products);
+        // Collapse products on colliding group labels
+        const collapsedProducts = values(groupBy(products, v => v.label)).map(t => ({
+          label: t[0].label,
+          assets: t.flatMap(v => v.assets),
+        }));
+
+        return presentBalanceFetcherResponse(collapsedProducts);
       }
     };
